Move UserDetailBok file
